Login finalizado, modulo nutricionista creado falta aprovacion, falta restringir los permisos, añadir filtros y optimizar

+++ api/models/Sale.js
@@ -25,7 +25,7 @@ module.exports = {
 			required: true
 		},
 		fullPrice: {
-			type :'float',
+			type :'decimal',
 			required: true
 		},
 		sellerName: {
